requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
importlib-metadata==0.19
minecart==0.3.0
more-itertools==7.2.0
packaging==19.1
pdfminer3k==1.3.1
Pillow==6.1.0
pluggy==0.12.0
ply==3.11
py==1.8.0
pyparsing==2.4.2
pytest==5.1.0
six==1.12.0
wcwidth==0.1.7
zipp==0.5.2
